ndt7: use int64 where it makes sense
It turns out the real reason why I was using float64 is that it
was unclear to me how to get an int64 from C code.

I also think it makes sense to use integers where we're dealing
with integer variables, because this is least surprising.

This diff basically rewrites the code that gathers BBR variables
to avoid using C code. The new implementation entails less source
files, and allows me to return directly a model.BBRInfo.

In turn, model.BBRInfo has an integer definition for the
MaxBandwidth, which is actually an integer.

I am using int64 rather than uint64 because Java doesn't have
uint64. At the same time, it should be noted that JSON does not
support any integer larger than 2^53.